doc-evaluator/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
# torch ставим в Dockerfile, тут можно опустить или оставить >=2.2.0
transformers>=4.36.0
sentence-transformers>=2.2.2
textstat==0.7.3
numpy
scikit-learn
# Новые либы для LLM
gigachat
google-generativeai
openai  # Используется и для ChatGPT, и для локальной Ollama
aiohttp # Для асинхронных запросов
pydantic-settings==2.1.0

# Test dependencies
pytest
pytest-asyncio
httpx==0.27.0
